+++ examples/metrics.py
@@ -5,11 +5,16 @@
 import os
 import platform
 import time
+from typing import Tuple
 
 import psutil
+from aleph_message.models import AlephMessage
 
+from aleph_client import AuthenticatedUserSession
 from aleph_client.chains.ethereum import get_fallback_account
-from aleph_client.synchronous import create_aggregate
+from aleph_client.conf import settings
+from aleph_client.types import MessageStatus
+from aleph_client.user_session import AuthenticatedUserSessionSync
 
 
 def get_sysinfo():
@@ -49,8 +54,10 @@ def get_cpu_cores():
     return [c._asdict() for c in psutil.cpu_times_percent(0, percpu=True)]
 
 
-def send_metrics(account, metrics):
-    return create_aggregate(account, "metrics", metrics, channel="SYSINFO")
+def send_metrics(
+    session: AuthenticatedUserSessionSync, metrics
+) -> Tuple[AlephMessage, MessageStatus]:
+    return session.create_aggregate(key="metrics", content=metrics, channel="SYSINFO")
 
 
 def collect_metrics():
@@ -64,11 +71,14 @@ def collect_metrics():
 
 def main():
     account = get_fallback_account()
-    while True:
-        metrics = collect_metrics()
-        message, status = send_metrics(account, metrics)
-        print("sent", message.item_hash)
-        time.sleep(10)
+    with AuthenticatedUserSession(
+        account=account, api_server=settings.API_HOST
+    ) as session:
+        while True:
+            metrics = collect_metrics()
+            message, status = send_metrics(session, metrics)
+            print("sent", message.item_hash)
+            time.sleep(10)
 
 
 if __name__ == "__main__":
