+++ examples/store.py
@@ -1,10 +1,8 @@
 import asyncio
 
-import aiohttp
 import click
 from aleph_message.models import StoreMessage
 
-from aleph_client.asynchronous import create_store
 from aleph_client.chains.common import get_fallback_private_key
 from aleph_client.chains.ethereum import ETHAccount
 from aleph_client.conf import settings
@@ -25,7 +23,9 @@ async def print_output_hash(message: StoreMessage, status: MessageStatus):
 
 
 async def do_upload(account, engine, channel, filename=None, file_hash=None):
-    async with AuthenticatedUserSession(account=account, api_server=settings.API_HOST) as session:
+    async with AuthenticatedUserSession(
+        account=account, api_server=settings.API_HOST
+    ) as session:
         print(filename, account.get_address())
         if filename:
             try:
@@ -35,8 +35,7 @@ async def do_upload(account, engine, channel, filename=None, file_hash=None):
                     if len(content) > 4 * 1024 * 1024 and engine == "STORAGE":
                         print("File too big for native STORAGE engine")
                         return
-                    message, status = await create_store(
-                        session=session,
+                    message, status = await session.create_store(
                         file_content=content,
                         channel=channel,
                         storage_engine=engine.lower(),
@@ -46,8 +45,7 @@ async def do_upload(account, engine, channel, filename=None, file_hash=None):
                 raise
 
         elif file_hash:
-            message, status = await create_store(
-                session=session,
+            message, status = await session.create_store(
                 file_hash=file_hash,
                 channel=channel,
                 storage_engine=engine.lower(),
