+++ src/aleph_client/__init__.py
@@ -1,4 +1,5 @@
 from pkg_resources import get_distribution, DistributionNotFound
+from .user_session import AuthenticatedUserSession, UserSession
 
 try:
     # Change here if project is renamed and does not equal the package name
@@ -8,3 +9,6 @@ except DistributionNotFound:
     __version__ = "unknown"
 finally:
     del get_distribution, DistributionNotFound
+
+
+__all__ = ["AuthenticatedUserSession", "UserSession"]
