+++ tests/integration/itest_aggregates.py
@@ -2,10 +2,6 @@ import pytest
 import json
 from typing import Dict
 
-from aleph_client.asynchronous import (
-    create_aggregate,
-    fetch_aggregate,
-)
 from aleph_client.user_session import AuthenticatedUserSession
 from tests.integration.toolkit import try_until
 from .config import REFERENCE_NODE, TARGET_NODE
@@ -21,9 +17,10 @@ async def create_aggregate_on_target(
     receiver_node: str,
     channel="INTEGRATION_TESTS",
 ):
-    async with AuthenticatedUserSession(account=account, api_server=emitter_node) as tx_session:
-        aggregate_message, message_status = await create_aggregate(
-            session=tx_session,
+    async with AuthenticatedUserSession(
+        account=account, api_server=emitter_node
+    ) as tx_session:
+        aggregate_message, message_status = await tx_session.create_aggregate(
             key=key,
             content=content,
             channel="INTEGRATION_TESTS",
@@ -40,11 +37,12 @@ async def create_aggregate_on_target(
     assert aggregate_message.content.address == account.get_address()
     assert aggregate_message.content.content == content
 
-    async with AuthenticatedUserSession(account=account, api_server=receiver_node) as rx_session:
+    async with AuthenticatedUserSession(
+        account=account, api_server=receiver_node
+    ) as rx_session:
         aggregate_from_receiver = await try_until(
-            fetch_aggregate,
+            rx_session.fetch_aggregate,
             lambda aggregate: aggregate is not None,
-            session=rx_session,
             timeout=5,
             address=account.get_address(),
             key=key,
