+++ tests/unit/test_synchronous_get.py
@@ -1,14 +1,12 @@
 from aleph_message.models import MessageType, MessagesResponse
 
 from aleph_client.conf import settings
-from aleph_client.synchronous import get_messages
 from aleph_client.user_session import UserSession
 
 
 def test_get_posts():
     with UserSession(api_server=settings.API_HOST) as session:
-        response: MessagesResponse = get_messages(
-            session=session,
+        response: MessagesResponse = session.get_messages(
             pagination=2,
             message_type=MessageType.post,
         )
